tests: clean up redundant tests
